RNR(ccems)	R Documentation

Ribonucleotide Reductase Data

Description

This is a compilation of ribonucleotide reductase (RNR) literature data into one data frame.

Usage

data(RNR)

Format

A data frame with 546 observations on the following 23 variables.

R

The total concentration of R1 (big R for the big subunit of RNR).

r

The total concentration of R2 dimer (small r for the small subunit of RNR).

U

The total concentration of UDP.

C

The total concentration of CDP.

G

The total concentration of GDP.

A

The total concentration of ADP.

t

The total concentration of dTTP.

tf

The concentration of free dTTP.

g

The total concentration of dGTP.

gf

The concentration of free dGTP.

a

The total concentration of dATP.

af

The concentration of free dATP.

X

The total concentration of ATP.

Mg

The total concentration of magnesium.

v

The velocity of the reduction reaction in 1/seconds; v in Scott et al should be multiplied by .00151 to create these units, save figure 7 (GDP reduction) which is in percentages of the maximum velocity.

m

The weight averaged average mass.

n

The number of ligands bound. Exception is Ingemarson et al. 1996 where it is a possibly negative BIAcore experiment parameter.

fg

The figure number in the original paper.

year

The year of the original paper.

jrnl

The journal of the data source: Bioc is Biochemistry, JBC is Journal of Biological Chemistry.

vol

The volume number of the article.

page

The page number of the article.

frstAut

The first author of the article. This is a factor with levels IngemarsonR, KashlanOB, Rofougaran, sanath, and ScottCP.

organism

An element of c("human","mouse","yeast","ecoli"). This dataset is predominantly mouse RNR data.

Details

All concentrations are in micromolar. To some extent, capital letters are used for ribonucleotides and small letters for deoxyribonucleotides. Free concentrations, where given, were model dependent and were thus converted back to total concentrations, i.e. to make them model independent. Values of zero that lead a sequence of measurements of increasing concentrations were set to 1e-2 (essentially zero) to allow [x]>0 subsetting.

Source

The figure number, year, first author, journal, volume and page number are all included in the dataframe. The main reference is Kashlan OB, Scott CP, Lear JD, Cooperman BS: A comprehensive model for the allosteric regulation of mammalian ribonucleotide reductase. Functional consequences of ATP- and dATP-induced oligomerization of the large subunit. Biochemistry 2002, 41(2):462-474. The first seven rows, by first author Sanath (Sanath Wijerathna of the Dealwis Lab at CWRU), are unpublished.

References

Radivoyevitch, T. (2008) Equilibrium model selection: dTTP induced R1 dimerization. BMC Systems Biology 2, 15.

Examples


library(ccems)
RNR[1:40,]
d=subset(RNR,(year==2002)&(fg==1)&(X>0),select=c(R,X,m,year))
plot(d$X,d$m,xlab="Total [ATP]", ylab="Weight averaged R1 mass", 
     main="Kashlan et al. Biochemistry, 2002, Fig. 1 (DLS data)")

[Package ccems version 1.01 Index]