mkEq(ccems)	R Documentation

Make Equality Constraints Lists

Description

This function converts a dataframe/chunk of models with either k or K equalities into a list structure that maps the trackers onto the independent parameter estimate.

Usage

 mkEq(g,chunk,activity=FALSE) 

Arguments

g 	 The output of mkg.

chunk 	 The chunk component of the output of mkGrids and mkSpurs.

activity 	 This is TRUE if the desired output is keq and FALSE if it is Keq.

Details

This function is already called internally by mkGrids and mkSpurs. Thus, the user should not need to call it. Note that the input activity can be false even if the dataset is activity data (i.e. activity in g is TRUE) since even then this function may be called to create K equalities rather than k equalities.

Value

Lists of either k or K equality constraints

Author(s)

Tom Radivoyevitch (txr24@case.edu)

References

Radivoyevitch, T. (2008) Equilibrium model selection: dTTP induced R1 dimerization. BMC Systems Biology 2, 15.

See Also

mkGrids, mkSpurs

[Package ccems version 1.01 Index]