mkHTML(ccems)	R Documentation

Make HTML File

Description

Writes the best models to an html summary file in the results directory.

Usage

 mkHTML(smodels,showConstr=FALSE) 

Arguments

smodels 	 A list of fitted model objects, preferably already sorted by best AIC.

showConstr 	 This is TRUE if constrained (fixed or tracking) parameters are to be included in the html output file. This is used for error checking.

Details

This function is already called internally by ems after each chunk of models is fitted. Thus, the user should not need to call it.

Value

No value. The output is an html file in the results directory.

Author(s)

Tom Radivoyevitch (txr24@case.edu)

References

Radivoyevitch, T. (2008) Equilibrium model selection: dTTP induced R1 dimerization. BMC Systems Biology 2, 15.

See Also

mkModel, ems

[Package ccems version 1.01 Index]