mkSingleThread(ccems)	R Documentation

Make Equality Constraints Lists

Description

This function creates k or K equality matrices for single threads.

Usage

 mkSingleThread(nNodes,g) 

Arguments

nNodes 	 The number of nodes in the thread.

g 	 The output of mkg. This is used only to grab the mylets field.

Details

This function is called internally by mkGrids and mkSpurs. Thus, the user should not need to call it.

Value

Matrix of characters of contiguous blocks of K or k equality constraints within a thread.

Author(s)

Tom Radivoyevitch (txr24@case.edu)

See Also

mkGrids, mkSpurs

[Package ccems version 1.01 Index]