testgC(ccems)	R Documentation

Test g Object TCC C Compilation

Description

This function runs lsoda using the compiled C code of the total concentration constraint (TCC) system equations that should have been created by mkg.

Usage

 testgC(g) 

Arguments

g 	 The output of mkg. Since mkg already calls this function, it has no independent use.

Details

Components of g used by this function are: g$parmsTCC which includes the total concentrations, the fraction of active protein p, and the spur graph complete dissociation constant parameter values Kj; and the lsoda ODE integration tolerance parameters g$rtol and g$atol.

Value

No value returned. Ten time points are printed to the screen. The free concentrations should show signs of convergence toward a steady state.

Author(s)

Tom Radivoyevitch

See Also

lsoda

Examples


library(ccems)
topology <- list(  
        heads=c("R1t0","R2t0"),  
        sites=list(       
                s=list(                     # s-site    thread #
                        m=c("R1t1"),        # monomer      1
                        d=c("R2t1","R2t2")  # dimer        2
                )
        )
) 
g=mkg(topology,TCC=TRUE) 
testgC(g) # this checks the compiled code again (it was already called in mkg) 

[Package ccems version 1.01 Index]